requirements.txt
numpy
torch
matplotlib
scikit-learn
seaborn
umap-learn
minisom
